+++ staging/src/k8s.io/pod-security-admission/cmd/webhook/server/server.go
@@ -24,6 +24,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"strconv"
 	"time"
 
 	"github.com/spf13/cobra"
@@ -37,11 +38,14 @@ import (
 	clientset "k8s.io/client-go/kubernetes"
 	restclient "k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
+	Version "k8s.io/component-base/version"
 	"k8s.io/klog/v2"
 	"k8s.io/pod-security-admission/admission"
 	admissionapi "k8s.io/pod-security-admission/admission/api"
 	podsecurityconfigloader "k8s.io/pod-security-admission/admission/api/load"
+	"k8s.io/pod-security-admission/api"
 	"k8s.io/pod-security-admission/cmd/webhook/server/options"
+	"k8s.io/pod-security-admission/metrics"
 	"k8s.io/pod-security-admission/policy"
 )
 
@@ -256,7 +260,7 @@ func Setup(c *Config) (*Server, error) {
 	s.delegate = &admission.Admission{
 		Configuration:    c.PodSecurityConfig,
 		Evaluator:        evaluator,
-		Metrics:          nil, // TODO: wire to default prometheus metrics
+		Metrics:          metrics.NewPrometheusRecorder(api.MajorMinorVersion(getAPIVersion())),
 		PodSpecExtractor: admission.DefaultPodSpecExtractor{},
 		PodLister:        admission.PodListerFromClient(client),
 		NamespaceGetter:  admission.NamespaceGetterFromListerAndClient(namespaceLister, client),
@@ -299,3 +303,19 @@ func parseTimeout(req *http.Request) (time.Duration, bool, error) {
 
 	return timeout, true, nil
 }
+
+// getAPIVersion get the version of apiServer and return the version major and minor
+func getAPIVersion() (major int, minor int) {
+	var err error
+	apiVersion := Version.Get()
+	major, err = strconv.Atoi(apiVersion.Major)
+	if err != nil {
+		return
+	}
+	minor, err = strconv.Atoi(apiVersion.Minor)
+	if err != nil {
+		return
+	}
+	return major, minor
+}
+
